# specs/001-terminal-designer-routing/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: LiquidTwin API (feature 001)
  version: 0.1.0
  description: Terminal management, validation, availability and quickest-route queries. No authentication in this feature.
servers:
  - url: /api/v1
paths:
...
  /terminals/{terminalId}/routes:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    post:
      summary: Quickest feasible routes for one operation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RouteRequest' }
      responses:
        '200':
          description: Routes (possibly empty) with explanations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /terminals/{terminalId}/optimization/prepare:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    post:
      summary: Prepare Stage 1 route candidates for an optimization scenario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizationScenario' }
      responses:
        '200':
          description: Candidate routes and blockers for every submitted job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationPreparation' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /terminals/{terminalId}/optimize:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    post:
      summary: Optimize a prepared multi-job scenario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OptimizationScenario' }
      responses:
        '200':
          description: Solver status, selected routes, KPIs, and timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OptimizationResult' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /terminals/{terminalId}/routes/confirm:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    post:
      summary: Confirm the final route chosen from the shortlist (FR-023)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request, route]
              properties:
                request: { $ref: '#/components/schemas/RouteRequest' }
                route: { $ref: '#/components/schemas/Route' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfirmedRoute' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /terminals/{terminalId}/routes/confirmed:
    parameters:
      - { $ref: '#/components/parameters/TerminalId' }
    get:
      summary: List confirmed routes (outdated ones flagged)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConfirmedRoute' }
...
components:
  parameters:
    TerminalId:
      name: terminalId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Input rejected; nothing was saved
      content:
        application/json:
          schema:
            type: object
            required: [issues]
            properties:
              issues:
                type: array
                items: { $ref: '#/components/schemas/ValidationIssue' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    ValidationIssue:
      type: object
      required: [severity, code, message]
      properties:
        severity: { type: string, enum: [ERROR, WARNING] }
        code: { type: string }
        element_id: { type: string }
        node_id: { type: string }
        row: { type: integer, description: CSV row for import errors }
        file: { type: string }
        message: { type: string }
        fix_hint: { type: string }
...
    RouteRequest:
      type: object
      required: [direction, product_id, volume_m3, rate_m3h, window_from, source_ids, destination_ids]
      properties:
        version: { type: integer, minimum: 1, description: Defaults to latest }
        direction: { type: string, enum: [IN, OUT, TRANSFER] }
        product_id: { type: string }
        volume_m3: { type: number, exclusiveMinimum: 0 }
        rate_m3h: { type: number, exclusiveMinimum: 0 }
        window_from: { type: string, format: date-time }
        window_to: { type: string, format: date-time, description: Defaults to window_from plus transfer time }
        source_ids: { type: array, minItems: 1, items: { type: string } }
        destination_ids: { type: array, minItems: 1, items: { type: string } }
        max_routes: { type: integer, minimum: 1, maximum: 20, default: 5 }
        pump_suction_inputs:
          type: array
          items:
            type: object
            required: [pump_id, npsh_available_m]
            properties:
              pump_id: { type: string }
              npsh_available_m: { type: number, minimum: 0 }
    OptimizationJob:
      type: object
      required: [id, direction, product_id, source_ids, destination_ids, volume_m3, rate_m3h, earliest_start_min, due_min]
      properties:
        id: { type: string, minLength: 1, maxLength: 64 }
        direction: { type: string, enum: [IN, OUT, TRANSFER] }
        product_id: { type: string }
        source_ids: { type: array, minItems: 1, maxItems: 100, items: { type: string } }
        destination_ids: { type: array, minItems: 1, maxItems: 100, items: { type: string } }
        volume_m3: { type: number, exclusiveMinimum: 0 }
        rate_m3h: { type: number, exclusiveMinimum: 0 }
        earliest_start_min: { type: integer, minimum: 0 }
        due_min: { type: integer, minimum: 0 }
        pump_suction_inputs:
          type: array
          items:
            type: object
            required: [pump_id, npsh_available_m]
            properties:
              pump_id: { type: string }
              npsh_available_m: { type: number, minimum: 0 }
    OptimizationScenario:
      type: object
      required: [terminal_version, horizon_start, jobs]
      properties:
        terminal_version: { type: integer, minimum: 1 }
        horizon_start: { type: string, format: date-time }
        horizon_minutes: { type: integer, minimum: 1, maximum: 10080, default: 1000 }
        jobs: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/OptimizationJob' } }
        objective_weights: { $ref: '#/components/schemas/ObjectiveWeights' }
        time_limit_seconds: { type: number, exclusiveMinimum: 0, maximum: 30, default: 30 }
        random_seed: { type: integer, minimum: 0, maximum: 2147483647, default: 1 }
    ObjectiveWeights:
      type: object
      properties:
        waiting: { type: integer, minimum: 0, maximum: 1000, default: 1 }
        lateness: { type: integer, minimum: 0, maximum: 1000, default: 20 }
        flush_volume: { type: integer, minimum: 0, maximum: 1000, default: 1 }
        valves: { type: integer, minimum: 0, maximum: 1000, default: 2 }
        shared_headers: { type: integer, minimum: 0, maximum: 1000, default: 3 }
    PreparedOptimizationJob:
      type: object
      required: [job_id, routes, blockers]
      properties:
        job_id: { type: string }
        routes: { type: array, items: { $ref: '#/components/schemas/Route' } }
        blockers: { type: array, items: { $ref: '#/components/schemas/Exclusion' } }
    OptimizationPreparation:
      type: object
      required: [terminal_version, horizon_start, horizon_minutes, engine_version, jobs]
      properties:
        terminal_version: { type: integer }
        horizon_start: { type: string, format: date-time }
        horizon_minutes: { type: integer }
        engine_version: { type: string }
        jobs: { type: array, items: { $ref: '#/components/schemas/PreparedOptimizationJob' } }
    OptimizedJob:
      type: object
      required: [job_id, start_min, end_min, late_min, route]
      properties:
        job_id: { type: string }
        start_min: { type: integer }
        end_min: { type: integer }
        late_min: { type: integer }
        route: { $ref: '#/components/schemas/Route' }
    UnscheduledJob:
      type: object
      required: [job_id, detail, blockers, conflict_elements]
      properties:
        job_id: { type: string }
        detail: { type: string }
        blockers: { type: array, items: { $ref: '#/components/schemas/Exclusion' } }
        conflict_elements: { type: array, items: { type: string } }
    ElementUseInterval:
      type: object
      required: [element_id, job_id, product_id, start_min, end_min]
      properties:
        element_id: { type: string }
        job_id: { type: string }
        product_id: { type: string }
        start_min: { type: integer }
        end_min: { type: integer }
    OptimizationKpis:
      type: object
      required: [on_time_jobs, total_lateness_min, waiting_min, flush_volume_m3, makespan_min]
      properties:
        on_time_jobs: { type: integer }
        total_lateness_min: { type: integer }
        waiting_min: { type: integer }
        flush_volume_m3: { type: integer }
        makespan_min: { type: integer }
    OptimizationResult:
      type: object
      required: [terminal_version, engine_version, solver_version, status, detail, scheduled_jobs, unscheduled_jobs, kpis, element_timeline]
      properties:
        terminal_version: { type: integer }
        engine_version: { type: string }
        solver_version: { type: string }
        status: { type: string, enum: [OPTIMAL, FEASIBLE, INFEASIBLE, NO_ROUTE, UNKNOWN] }
        detail: { type: string }
        objective: { type: [integer, 'null'] }
        best_bound: { type: [integer, 'null'] }
        scheduled_jobs: { type: array, items: { $ref: '#/components/schemas/OptimizedJob' } }
        unscheduled_jobs: { type: array, items: { $ref: '#/components/schemas/UnscheduledJob' } }
        kpis: { $ref: '#/components/schemas/OptimizationKpis' }
        element_timeline: { type: array, items: { $ref: '#/components/schemas/ElementUseInterval' } }
    RouteResponse:
      type: object
      required: [terminal_version, engine_version, routes, exclusions]
      properties:
        terminal_version: { type: integer }
        engine_version: { type: string }
        routes:
          type: array
          items: { $ref: '#/components/schemas/Route' }
        exclusions:
          type: array
          items: { $ref: '#/components/schemas/Exclusion' }
        no_route:
          $ref: '#/components/schemas/NoRoute'
    Route:
      type: object
      required: [rank, source_id, destination_id, steps, metrics]
      properties:
        rank: { type: integer }
        source_id: { type: string }
        destination_id: { type: string }
        steps:
          type: array
          items:
            type: object
            required: [element_id, from, to, reversed]
            properties:
              element_id: { type: string }
              from: { type: string }
              to: { type: string }
              reversed: { type: boolean }
        metrics:
          type: object
          required: [fill_min, transfer_min, flush_min, flush_volume_m3, total_min, valves, common_headers, head_margin_m, max_velocity_ms]
          properties:
            fill_min: { type: number }
            transfer_min: { type: number }
            flush_min: { type: number }
            flush_volume_m3: { type: number }
            total_min: { type: number }
            valves: { type: integer }
            common_headers: { type: integer }
            head_margin_m: { type: number }
            max_velocity_ms: { type: number }
            operating_flow_m3h: { type: [number, 'null'] }
            pump_head_m: { type: [number, 'null'] }
            system_head_m: { type: [number, 'null'] }
            suction_margin_m: { type: [number, 'null'] }
            pump_speed_ratio: { type: [number, 'null'] }
    Exclusion:
      type: object
      required: [reason, detail]
      properties:
        element_id: { type: string }
        node_id: { type: string }
        reason:
          type: string
          enum: [NOT_CERTIFIED, VELOCITY, RESIDUE, ONE_WAY_PUMP, UNAVAILABLE, DEDICATED_OTHER_GROUP, PUMP_HEAD, ENDPOINT_STOCK, ENDPOINT_SPACE, ENDPOINT_PRODUCT, ENDPOINT_DIRECTION, INVALID_PUMP_CURVE, PUMP_FLOW_OUT_OF_RANGE, NO_PUMP_SYSTEM_INTERSECTION, PUMP_SPEED_OUT_OF_RANGE, PUMP_SUCTION_MARGIN]
        detail: { type: string }
    ConfirmedRoute:
      type: object
      required: [id, terminal_version, request, route, outdated, created_at]
      properties:
        id: { type: string, format: uuid }
        terminal_version: { type: integer }
        request: { $ref: '#/components/schemas/RouteRequest' }
        route: { $ref: '#/components/schemas/Route' }
        outdated: { type: boolean }
        outdated_reason: { type: string }
        created_at: { type: string, format: date-time }
    NoRoute:
      type: object
      required: [message, blockers]
      properties:
        message: { type: string }
        blockers:
          type: array
          items: { $ref: '#/components/schemas/Exclusion' }